# cookies: improve validateCookieValue