--- a/benchmarks/cookies/validate-cookie-value.mjs
+++ b/benchmarks/cookies/validate-cookie-value.mjs
@@ -1,0 +1,16 @@
+import { bench, group, run } from 'mitata'
+import { validateCookieValue } from '../../lib/web/cookies/util.js'
+
+const valid = 'Cat'
+const wrappedValid = `"${valid}"`
+
+group('validateCookieValue', () => {
+  bench(`valid: ${valid}`, () => {
+    return validateCookieValue(valid)
+  })
+  bench(`valid: ${wrappedValid}`, () => {
+    return validateCookieValue(wrappedValid)
+  })
+})
+
+await run()