--- a/lib/web/cookies/util.js
+++ b/lib/web/cookies/util.js
@@ -69,18 +69,30 @@ function validateCookieName (name) {
  * @param {string} value
  */
 function validateCookieValue (value) {
-  for (const char of value) {
-    const code = char.charCodeAt(0)
+  let len = value.length
+  let i = 0
+
+  // if the value is wrapped in DQUOTE
+  if (value[0] === '"') {
+    if (len === 1 || value[len - 1] !== '"') {
+      throw new Error('Invalid cookie value')
+    }
+    --len
+    ++i
+  }
+
+  while (i < len) {
+    const code = value.charCodeAt(i++)
 
     if (
       code < 0x21 || // exclude CTLs (0-31)
-      code === 0x22 ||
-      code === 0x2C ||
-      code === 0x3B ||
-      code === 0x5C ||
-      code > 0x7E // non-ascii
+      code > 0x7E || // non-ascii and DEL (127)
+      code === 0x22 || // "
+      code === 0x2C || // ,
+      code === 0x3B || // ;
+      code === 0x5C // \
     ) {
-      throw new Error('Invalid header value')
+      throw new Error('Invalid cookie value')
     }
   }
 }
@@ -286,6 +298,7 @@ function getHeadersList (headers) {
 module.exports = {
   isCTLExcludingHtab,
   validateCookiePath,
+  validateCookieValue,
   toIMFDate,
   stringify,
   getHeadersList