- fixes it to be conforming with the spec allowing wrapped values in double quotes
- improve performance
- fix error message


before:
```sh
aras@aras-Lenovo-Legion-5-17ARH05H:~/workspace/undici$ node benchmarks/cookies/validate-cookie-value.mjs 
cpu: AMD Ryzen 7 4800H with Radeon Graphics
runtime: node v21.6.2 (x64-linux)

benchmark         time (avg)             (min … max)       p75       p99      p999
---------------------------------------------------- -----------------------------
• validateCookieValue
---------------------------------------------------- -----------------------------
valid: Cat     19.97 ns/iter     (18.45 ns … 656 ns)   19.4 ns  31.65 ns  38.36 ns
valid: "Cat"  error: Invalid header value
Error: Invalid header value
    at validateCookieValue (/home/aras/workspace/undici/lib/web/cookies/util.js:83:13)
    at file:///home/aras/workspace/undici/benchmarks/cookies/validate-cookie-value.mjs:12:12
    at measure (file:///home/aras/workspace/undici/benchmarks/node_modules/mitata/src/lib.mjs:21:15)
    at run (file:///home/aras/workspace/undici/benchmarks/node_modules/mitata/src/cli.mjs:270:28)
    at async file:///home/aras/workspace/undici/benchmarks/cookies/validate-cookie-value.mjs:16:1

summary for validateCookieValue
  valid: Cat
```

after:

```sh
aras@aras-Lenovo-Legion-5-17ARH05H:~/workspace/undici$ node benchmarks/cookies/validate-cookie-value.mjs 
cpu: AMD Ryzen 7 4800H with Radeon Graphics
runtime: node v21.6.2 (x64-linux)

benchmark         time (avg)             (min … max)       p75       p99      p999
---------------------------------------------------- -----------------------------
• validateCookieValue
---------------------------------------------------- -----------------------------
valid: Cat     11.64 ns/iter      (9.41 ns … 599 ns)  10.88 ns  43.89 ns  72.91 ns
valid: "Cat"   12.98 ns/iter      (11.8 ns … 231 ns)  12.96 ns  17.26 ns  18.82 ns

summary for validateCookieValue
  valid: Cat
   1.11x faster than valid: "Cat"
```